+++ src/Tests/Infrastructure/New-ModuleProject.Tests.ps1
@@ -97,6 +97,13 @@ Describe 'Module Infra Tests' {
                     $moduleOnlyFiles.Name.Contains('bug-report.md') | Should -BeExactly $false
                     $moduleOnlyFiles.Name.Contains('feature_request.md') | Should -BeExactly $false
 
+                    # ReadtheDocs
+                    $moduleOnlyFiles.Name.Contains('mkdocs.yml') | Should -BeExactly $false
+                    $moduleOnlyFiles.Name.Contains('requirements.txt') | Should -BeExactly $false
+                    $moduleOnlyFiles.Name.Contains('.readthedocs.yaml') | Should -BeExactly $false
+                    $moduleOnlyFiles.Name.Contains('index.md') | Should -BeExactly $false
+
+
                     # AWS
                     $moduleOnlyFiles.Name.Contains('buildspec_powershell_windows.yml') | Should -BeExactly $false
                     $moduleOnlyFiles.Name.Contains('buildspec_pwsh_linux.yml') | Should -BeExactly $false
@@ -145,6 +152,7 @@ Describe 'Module Infra Tests' {
                         CICD        = 'CODEBUILD'
                         AWSOptions  = 'ps', 'pwshcore', 'pwsh'
                         RepoType    = 'GITHUB'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -203,6 +211,7 @@ Describe 'Module Infra Tests' {
                         CICD        = 'CODEBUILD'
                         AWSOptions  = 'ps', 'pwshcore', 'pwsh'
                         RepoType    = 'BITBUCKET'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -261,6 +270,7 @@ Describe 'Module Infra Tests' {
                         CICD        = 'CODEBUILD'
                         AWSOptions  = 'ps', 'pwshcore', 'pwsh'
                         RepoType    = 'CodeCommit'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -309,6 +319,7 @@ Describe 'Module Infra Tests' {
                         CICD        = 'CODEBUILD'
                         AWSOptions  = 'ps', 'pwshcore', 'pwsh'
                         RepoType    = 'AZURE'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -349,6 +360,7 @@ Describe 'Module Infra Tests' {
                         CICD         = 'AZURE'
                         AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
                         RepoType     = 'GITHUB'
+                        ReadtheDocs  = 'None'
                         License      = 'None'
                         Changelog    = 'NOCHANGELOG'
                         COC          = 'NOCONDUCT'
@@ -387,6 +399,7 @@ Describe 'Module Infra Tests' {
                         CICD         = 'AZURE'
                         AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
                         RepoType     = 'AZURE'
+                        ReadtheDocs  = 'None'
                         License      = 'None'
                         Changelog    = 'NOCHANGELOG'
                         COC          = 'NOCONDUCT'
@@ -425,6 +438,7 @@ Describe 'Module Infra Tests' {
                         CICD         = 'AZURE'
                         AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
                         RepoType     = 'BITBUCKET'
+                        ReadtheDocs  = 'None'
                         License      = 'None'
                         Changelog    = 'NOCHANGELOG'
                         COC          = 'NOCONDUCT'
@@ -467,6 +481,7 @@ Describe 'Module Infra Tests' {
                         CICD            = 'APPVEYOR'
                         AppveyorOptions = 'windows', 'pwshcore', 'linux', 'macos'
                         RepoType        = 'GITHUB'
+                        ReadtheDocs     = 'None'
                         License         = 'None'
                         Changelog       = 'NOCHANGELOG'
                         COC             = 'NOCONDUCT'
@@ -508,6 +523,7 @@ Describe 'Module Infra Tests' {
                         CICD           = 'GITHUB'
                         GitHubAOptions = 'windows', 'pwshcore', 'linux', 'macos'
                         RepoType       = 'GITHUB'
+                        ReadtheDocs    = 'None'
                         License        = 'None'
                         Changelog      = 'NOCHANGELOG'
                         COC            = 'NOCONDUCT'
@@ -560,6 +576,7 @@ Describe 'Module Infra Tests' {
                         FN          = 'user full name'
                         CICD        = 'BITBUCKET'
                         RepoType    = 'BITBUCKET'
+                        ReadtheDocs = 'None'
                         License     = 'None'
                         Changelog   = 'NOCHANGELOG'
                         COC         = 'NOCONDUCT'
@@ -598,6 +615,7 @@ Describe 'Module Infra Tests' {
                         CICD          = 'GITLAB'
                         GitLabOptions = 'windows', 'pwshcore', 'linux'
                         RepoType      = 'GITLAB'
+                        ReadtheDocs   = 'None'
                         License       = 'None'
                         Changelog     = 'NOCHANGELOG'
                         COC           = 'NOCONDUCT'
@@ -642,6 +660,7 @@ Describe 'Module Infra Tests' {
                         FN          = 'user full name'
                         CICD        = 'NONE'
                         RepoType    = 'CODECOMMIT'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -696,6 +715,7 @@ Describe 'Module Infra Tests' {
                         FN          = 'user full name'
                         CICD        = 'NONE'
                         RepoType    = 'GITHUB'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -771,6 +791,7 @@ Describe 'Module Infra Tests' {
                         FN          = 'user full name'
                         CICD        = 'NONE'
                         RepoType    = 'GITLAB'
+                        ReadtheDocs = 'None'
                         License     = 'MIT'
                         Changelog   = 'CHANGELOG'
                         COC         = 'CONDUCT'
@@ -834,6 +855,99 @@ Describe 'Module Infra Tests' {
 
         } #context_repo
 
+        Context 'Read the Docs' {
+
+            It 'should generate the appropriate files for the standard readthedocs theme' {
+                $moduleParameters = @{
+                    VAULT       = 'text'
+                    ModuleName  = 'modulename'
+                    Description = 'text'
+                    Version     = '0.0.1'
+                    FN          = 'user full name'
+                    CICD        = 'NONE'
+                    RepoType    = 'GITHUB'
+                    ReadtheDocs = 'READTHEDOCS'
+                    RTDTheme    = 'READTHEDOCSTHEME'
+                    License     = 'None'
+                    Changelog   = 'NOCHANGELOG'
+                    COC         = 'NOCONDUCT'
+                    Contribute  = 'NOCONTRIBUTING'
+                    Security    = 'NOSECURITY'
+                    CodingStyle = 'Stroustrup'
+                    Help        = 'Yes'
+                    Pester      = '5'
+                    S3Bucket    = 'PSGallery'
+                    PassThru    = $true
+                    NoLogo      = $true
+                }
+                $eval = New-ModuleProject -ModuleParameters $moduleParameters -DestinationPath $outPutPath
+                $eval | Should -Not -BeNullOrEmpty
+
+                $rtdFiles = Get-ChildItem -Path $outPutPathStar -Recurse -Force
+
+                # read the docs
+                $rtdFiles.Name.Contains('mkdocs.yml') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('requirements.txt') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('.readthedocs.yaml') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('index.md') | Should -BeExactly $true
+
+                $mkdocsContentPath = [System.IO.Path]::Combine($outPutPath, 'mkdocs.yml')
+                $mkdocsContent = Get-Content -Path $mkdocsContentPath -Raw
+                $mkdocsContent | Should -Not -BeLike '*material*'
+                $mkdocsContent | Should -BeLike '*modulename*'
+
+                $reqContentPath = [System.IO.Path]::Combine($outPutPath, 'docs', 'requirements.txt')
+                $reqContent = Get-Content -Path $reqContentPath -Raw
+                $reqContent | Should -Not -BeLike '*material*'
+
+            } #it
+
+            It 'should generate the appropriate files for the material theme' {
+                $moduleParameters = @{
+                    VAULT       = 'text'
+                    ModuleName  = 'modulename'
+                    Description = 'text'
+                    Version     = '0.0.1'
+                    FN          = 'user full name'
+                    CICD        = 'NONE'
+                    RepoType    = 'GITHUB'
+                    ReadtheDocs = 'READTHEDOCS'
+                    RTDTheme    = 'MATERIALTHEME'
+                    License     = 'None'
+                    Changelog   = 'NOCHANGELOG'
+                    COC         = 'NOCONDUCT'
+                    Contribute  = 'NOCONTRIBUTING'
+                    Security    = 'NOSECURITY'
+                    CodingStyle = 'Stroustrup'
+                    Help        = 'Yes'
+                    Pester      = '5'
+                    S3Bucket    = 'PSGallery'
+                    PassThru    = $true
+                    NoLogo      = $true
+                }
+                $eval = New-ModuleProject -ModuleParameters $moduleParameters -DestinationPath $outPutPath
+                $eval | Should -Not -BeNullOrEmpty
+
+                $rtdFiles = Get-ChildItem -Path $outPutPathStar -Recurse -Force
+
+                # read the docs
+                $rtdFiles.Name.Contains('mkdocs.yml') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('requirements.txt') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('.readthedocs.yaml') | Should -BeExactly $true
+                $rtdFiles.Name.Contains('index.md') | Should -BeExactly $true
+
+                $mkdocsContentPath = [System.IO.Path]::Combine($outPutPath, 'mkdocs.yml')
+                $mkdocsContent = Get-Content -Path $mkdocsContentPath -Raw
+                $mkdocsContent | Should -BeLike '*material*'
+
+                $reqContentPath = [System.IO.Path]::Combine($outPutPath, 'docs', 'requirements.txt')
+                $reqContent = Get-Content -Path $reqContentPath -Raw
+                $reqContent | Should -BeLike '*material*'
+
+            } #it
+
+        } #context_readthedocs
+
         Context 'Help Examples' {
 
             It 'should have a working example for vanilla module' {
@@ -865,6 +979,7 @@ Describe 'Module Infra Tests' {
                     CICD           = 'GITHUB'
                     GitHubAOptions = 'windows', 'pwshcore', 'linux', 'macos'
                     RepoType       = 'GITHUB'
+                    ReadtheDocs    = 'None'
                     License        = 'MIT'
                     Changelog      = 'CHANGELOG'
                     COC            = 'CONDUCT'
@@ -892,6 +1007,7 @@ Describe 'Module Infra Tests' {
                     CICD        = 'CODEBUILD'
                     AWSOptions  = 'ps', 'pwshcore', 'pwsh'
                     RepoType    = 'GITHUB'
+                    ReadtheDocs = 'None'
                     License     = 'MIT'
                     Changelog   = 'CHANGELOG'
                     COC         = 'CONDUCT'
@@ -919,6 +1035,7 @@ Describe 'Module Infra Tests' {
                     CICD         = 'AZURE'
                     AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
                     RepoType     = 'GITHUB'
+                    ReadtheDocs  = 'None'
                     License      = 'None'
                     Changelog    = 'NOCHANGELOG'
                     COC          = 'NOCONDUCT'
@@ -945,6 +1062,7 @@ Describe 'Module Infra Tests' {
                     CICD            = 'APPVEYOR'
                     AppveyorOptions = 'windows', 'pwshcore', 'linux', 'macos'
                     RepoType        = 'GITHUB'
+                    ReadtheDocs     = 'None'
                     License         = 'None'
                     Changelog       = 'NOCHANGELOG'
                     COC             = 'NOCONDUCT'
@@ -962,7 +1080,8 @@ Describe 'Module Infra Tests' {
                 $manifestContent = Get-Content -Path $manifestContentPath -Raw
                 $manifestContent | Should -BeLike '*My awesome module is awesome*'
             } #it
-        }
+
+        } #context_help_examples
 
     } #context_module_checks